Fix -o insert index, support a.out default name
silly-cc won't work properly with Autoconf's ./configure scripts unless
it supports compiling without a "-o" argument.  Clang doesn't generate
an "a.out" file if "-o" is omitted, but "-MJ" is specified, so we have
to add "-o" and "a.out" arguments to our silly_args that get passed to
execv().